+++ src/apis/dlog/info.py
@@ -15,7 +15,7 @@ async def get_list(request: Request, response: Response, authorization: str = He
         page: Optional[int] = 1, page_size: Optional[int] = 10, \
         order_by: Optional[str] = "logid", order: Optional[str] = "desc", \
         speed_limit: Optional[int] = None, userid: Optional[int] = None, \
-        start_time: Optional[int] = None, end_time: Optional[int] = None, game: Optional[int] = None, status: Optional[int] = 1,\
+        after: Optional[int] = None, before: Optional[int] = None, game: Optional[int] = None, status: Optional[int] = 1,\
         challenge: Optional[str] = "any", division: Optional[str] = "any"):
     app = request.app
     dhrid = request.state.dhrid
@@ -84,8 +84,8 @@ async def get_list(request: Request, response: Response, authorization: str = He
             pass
     
     timelimit = ""
-    if start_time is not None and end_time is not None:
-        timelimit = f"AND dlog.timestamp >= {start_time} AND dlog.timestamp <= {end_time}"
+    if after is not None and before is not None:
+        timelimit = f"AND dlog.timestamp >= {after} AND dlog.timestamp <= {before}"
     
     if speed_limit is not None:
         speed_limit = f" AND dlog.topspeed <= {speed_limit}"